Rename completed
