+++ autoform/_wizard.js
@@ -4,7 +4,7 @@ var r = {};
 var cr = require('../config.json').resources;
 var k = require('../config.json').plugins;
 var b = require('../config.json').APIVARS.PLUGINS;
-var n = __dirname.replace('/core', b.DIR);
+var n = __dirname.replace('/autoform', b.DIR);
 var w = ['tokens', 'settings'];
 var s = ['actions', 'login', 'recover', 'register' ];
 var p = './views/index/forms';
