+++ views/browser/index.jade
@@ -1,0 +1,14 @@
+html
+  head
+  body
+    title Old Browser
+    div(style='text-align: center')
+      h1 Old browser detected
+      div
+        p Please download a modern browser
+      div
+        a(href='https://www.google.com/chrome') Google Chrome
+      div
+        a(href='https://www.mozilla.org') Mozilla Firefox
+      div
+        a(href='http://windows.microsoft.com/en-us/internet-explorer/download-ie') Internet Explorer
