+++ web/routes.js
@@ -15,6 +15,16 @@ web.set('views', views);
 web.set('view engine', 'jade');
 web.use(express.static(__dirname.replace('/web', '/public')));
 
+// old browser
+web.all('*', function(req, res, next) {
+  var regex = /(MSIE [1-3].0|Mozilla\/4.0)/g;
+  var header = req.headers['user-agent'].match(regex);
+  if(header && header.length){
+    return res.render('browser');
+  }
+  next();
+});
+
 // resources
 for(var route in resources){
   if(resources.hasOwnProperty(route)){
